duy OK phần class và tạo deck cho class

+++ src/backend/src/services/class.service.js
@@ -1,20 +1,20 @@
 'use strict'
 
 const { pick } = require("lodash")
-const { NotFoundError } = require("../core/error.response")
-const { ClassModel, ClassMember } = require("../models")
+const { NotFoundError, BadRequestError } = require("../core/error.response")
+const { ClassModel, ClassMember, User } = require("../models")
 const { createUniqueInviteCode } = require("../utils")
  
 class ClassService {
     static createClass = async({name, description, userId }) => {
         console.log("name, description, userId::", name, description, userId)
         if(!name || !description || !userId) {
-            throw new NotFoundError("Name, description or userId is required")
+            throw new BadRequestError("Name, description or userId is required")
         }
         
         const inviteCode = await createUniqueInviteCode();
         if(!inviteCode)
-            throw new NotFoundError("create invite code failed")
+            throw new BadRequestError("create invite code failed")
 
          const newClass = await ClassModel.create({
             name: name,
@@ -24,7 +24,7 @@ class ClassService {
             member_count: 1,
         })
         if(!newClass)
-            throw new NotFoundError("Create class failed")
+            throw new BadRequestError("Create class failed")
 
         const newMember = await ClassMember.create({
             user_id: userId,
@@ -46,18 +46,53 @@ class ClassService {
     static joinClass = async({userId, codeInvite}) => {
         const foundClass = await ClassModel.findOne({where: {code_invite: codeInvite}})
         if(!foundClass)
-            return new NotFoundError("Class not found")
+            throw new NotFoundError("Class not found")
 
-        const newMember = await MemberModel.create({userId, classId: foundClass.id})
-        if(!newMember)
-            return new NotFoundError("Join class failed")
+        const joinedMember = await ClassMember.findOne({where: {user_id: userId, class_id: foundClass.id}})
+        if(joinedMember)
+            throw new BadRequestError("You have already joined this class")
 
-        return "Join class successfully"
+        const newMember = await ClassMember.create({user_id: userId, class_id: foundClass.id})
+        const updatedClass = await ClassModel.update({member_count: foundClass.member_count + 1}, {where: {id: foundClass.id}})
+        if(!newMember || !updatedClass)
+            throw new BadRequestError("Join class failed")
+        return {
+            member: pick(newMember, ["id", "user_id", "class_id", "role"]),
+        }
     }
-
+   
     static getMembers = async({classId}) => {
-        const members = await ClassMember.findAll({where: {group_id: classId}})
-        return members
+        const members = await ClassMember.findAll({
+            where: { class_id: classId },
+            include: [{
+                model: User,
+                as: "User",
+                attributes: ['id', 'name', 'email']
+            }],
+            order: [
+                ['joined_at', 'ASC']
+            ]   
+        });
+
+        // làm phẳng data
+        const flattenedMembers = members.map(member => ({
+            memberId: member.id,
+            userId: member.User.id,
+            userName: member.User.name,
+            userEmail: member.User.email,
+            classId: member.class_id,
+            role: member.role,
+            joinedAt: member.joined_at
+        }));
+
+        return {
+            memberCount: flattenedMembers.length,
+            members: flattenedMembers
+        };
+    }
+    static getAllClasses = async() => {
+        const classes = await ClassModel.findAll()
+        return classes
     }
     static getClassesByUserId = async({userId, sortBy = "member_count"}) => {
         const order = [];
@@ -69,8 +104,7 @@ class ClassService {
         } else if (sortBy === 'updated_at') {
             order.push([{ model: ClassModel }, 'updated_at', 'DESC']);
         }
-        console.log(`userId: ${userId}`)
-        console.log(`order: ${order}`)
+
         const classes = await ClassMember.findAll({
             where: { user_id: userId },
             include: [{
@@ -83,21 +117,39 @@ class ClassService {
 
         return classes
     }
-    static deleteClass = async({classId}) => {
-        const deletedClass = await ClassModel.destroy({where: {id: classId}})
-        return deletedClass
-    }
     static updateClass = async({classId, name, description}) => {
+        if(!name || !description)
+            throw new BadRequestError("Name or description is required")
         const updatedClass = await ClassModel.update({name, description}, {where: {id: classId}})
         return updatedClass
     }
-    static deleteMember = async({userId, classId}) => {
-        const deletedMember = await ClassMember.destroy({where: {user_id: userId, group_id: classId}})
+    static deleteClass = async({classId, userId}) => {
+        const deletedClass = await ClassModel.destroy({where: {id: classId, host_user_id: userId}})
+        if(!deletedClass)
+            throw new BadRequestError("Delete class failed. You are not the host of this class")
+        return deletedClass
+    }
+    static deleteMember = async({userId, classId, hostUserId}) => { 
+        const classInfo = await ClassModel.findOne({ where: { id: classId, host_user_id: userId } });
+        if (classInfo) {
+            throw new BadRequestError("Member cannot delete member");
+        }
+
+        const deletedMember = await ClassMember.destroy({where: {user_id: userId, class_id: classId}})
+        if(!deletedMember)
+            throw new BadRequestError("Delete member failed")
         return deletedMember
     }
-    static updateMember = async({userId, classId, role}) => {
-        const updatedMember = await ClassMember.update({role}, {where: {user_id: userId, group_id: classId}})
-        return updatedMember
+    static leaveClass = async({userId, classId}) => {
+        const classInfo = await ClassModel.findOne({ where: { id: classId, host_user_id: userId } });
+        if (classInfo) {
+            throw new BadRequestError("Host cannot leave the class");
+        }
+
+        const deletedMember = await ClassMember.destroy({ where: { user_id: userId, class_id: classId } });
+        if(!deletedMember)
+            throw new BadRequestError("Leave class failed")
+        return deletedMember;
     }
 }
